requirements.txt
# Core Dependencies
openai==1.54.0
python-dotenv==1.0.0

# Vector Database
qdrant-client==1.11.3

# Document Processing
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.5
pypdf==5.1.0
python-docx==1.1.2

# Data Processing
numpy==1.26.4
pandas==2.2.3

# YouTube Transcript
youtube-transcript-api==0.6.2

# Web Framework
fastapi==0.115.4
uvicorn==0.32.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Utilities
tiktoken==0.8.0
beautifulsoup4==4.12.3
requests==2.32.3
